+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/manual/searchtag/TagAutocompleteDispatcher.java
@@ -16,20 +16,21 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core.analysis.manual.segment;
+package org.apache.skywalking.oap.server.core.analysis.manual.searchtag;
 
 import org.apache.skywalking.oap.server.core.analysis.SourceDispatcher;
 import org.apache.skywalking.oap.server.core.analysis.worker.MetricsStreamProcessor;
-import org.apache.skywalking.oap.server.core.source.TraceTagAutocomplete;
+import org.apache.skywalking.oap.server.core.source.TagAutocomplete;
 
-public class TraceTagAutocompleteDispatcher implements SourceDispatcher<TraceTagAutocomplete> {
+public class TagAutocompleteDispatcher implements SourceDispatcher<TagAutocomplete> {
 
     @Override
-    public void dispatch(TraceTagAutocomplete source) {
-        TraceTagAutocompleteData autocomplete = new TraceTagAutocompleteData();
+    public void dispatch(TagAutocomplete source) {
+        TagAutocompleteData autocomplete = new TagAutocompleteData();
         autocomplete.setTag(source.getTag());
         autocomplete.setTagKey(source.getTagKey());
         autocomplete.setTagValue(source.getTagValue());
+        autocomplete.setTagType(source.getTagType().name());
         autocomplete.setTimeBucket(source.getTimeBucket());
         MetricsStreamProcessor.getInstance().in(autocomplete);
     }