+++ oap-server/server-receiver-plugin/zipkin-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zipkin/trace/SpanForward.java
@@ -24,7 +24,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.apache.skywalking.oap.server.core.Const;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
-import org.apache.skywalking.oap.server.core.source.TraceTagAutocomplete;
+import org.apache.skywalking.oap.server.core.analysis.manual.searchtag.TagType;
+import org.apache.skywalking.oap.server.core.source.TagAutocomplete;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
@@ -111,10 +112,11 @@ public class SpanForward {
     }
 
     private void addAutocompleteTags(final long minuteTimeBucket, final String key, final String value, final String tagString) {
-        TraceTagAutocomplete tagAutocomplete = new TraceTagAutocomplete();
+        TagAutocomplete tagAutocomplete = new TagAutocomplete();
         tagAutocomplete.setTag(tagString);
         tagAutocomplete.setTagKey(key);
         tagAutocomplete.setTagValue(value);
+        tagAutocomplete.setTagType(TagType.TRACE);
         tagAutocomplete.setTimeBucket(minuteTimeBucket);
         receiver.receive(tagAutocomplete);
     }